refactor(turns): decompose run_profile/host.rs into a host/ directory module
`host.rs` had grown to ~2,592 lines — the largest file in the crate, well past
the 1,500 threshold (it carried an `arch-exempt: large_file`). It mixed ~11
independent host-contract clusters. Split it, move-only, into `host/`:

- mod.rs (62) — module decls + `pub use` re-exports (public `run_profile::X`
  paths unchanged)
- capability.rs (753), model.rs (334), refs.rs (322, both macros + all ten
  bounded-ref newtypes), validate.rs (305), progress.rs (245),
  run_context.rs (224), error.rs (157), context.rs (137), checkpoint.rs (120),
  transcript.rs (67), input.rs (63)

Move-only: bodies unchanged; a few former module-private helpers were widened to
`pub(crate)` so sibling submodules reach them (none were ever re-exported, so
the public API is identical). No file now exceeds 1,500 lines; the
`arch-exempt: large_file` marker is dropped.

Also `cargo fmt`s the crate, absorbing formatting drift left by the earlier
dedup commits in `coordinator.rs` / `lifecycle.rs` / `snapshot.rs`.

ironclaw_turns suite (15 binaries, 607 lib tests) + clippy --all-targets
--all-features + fmt --check green.
